requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
datasets==2.9.0
dill==0.3.6
Django==4.1.5
django-crispy-forms==1.14.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.12.0
idna==3.4
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pyarrow==10.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
responses==0.18.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.1.0
tldextract==3.4.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
xxhash==3.2.0
yarl==1.8.2
